+++ queueio/django/djp.py
@@ -1,0 +1,6 @@
+import djp
+
+
+@djp.hookimpl
+def installed_apps():
+    return ["queueio.django"]