+++ queueio/samples/django/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class QueueIOSampleConfig(AppConfig):
+    name = "queueio.samples.django"
+    label = "queueio_sample"