+++ queueio/samples/django/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .job import index
+
+urlpatterns = [
+    path("", index),
+]